feat: refactor Marketplace and Dashboard components to utilize global state for product listings and loading state management
